anothere change

+++ src/Graph.h
@@ -26,7 +26,7 @@ private:
 	void add(string element);
 	vector<string> reconstruct_path(int start, int goal, unordered_map<int, int> came_from);
 public:
-	unordered_map <string, pair<int, int>> airportCoordinates;
+	unordered_map <string, pair<double,double>> airportCoordinates;
 	void insert(string from, string to, int weight);
 	vector<string> findShortestPath_BFS(string from, string to);
 	std::pair<vector<string>, double> findShortestPath_Astar(string from, string to);
